Scriptable object usage way adjust
Update:
1. Because of some mistakes, have to adjust the way to use
scriptable object, so change some scriptable object usage.
such as playerstatus, equipmentslot, cardlist...
Now it should work like this:
start game and load value from the SO to a controller script,
and all the changes will happen in this script ,
when you want to save or close, then save the
values in the script into SO or a Json file.
2. Add drop_exp, drop_card to monster SO, which will
drop to player when it dead.
3. Add Levelupsystem to player, and you can now level
up due to kill monster.