Create player interact system
Create a villiage npc that player can interact with.
about interact system, create 3 type of interact template scriptable object: conversation, mission and transaction.
each of these template and set difference type of interact,
but now it only finish the conversation type, which is the most easier.
Also, to drive the interact, create interact controller and it's UI controller, which respectively responsible for interact system and it's display.